+++ tools/typesupport/tests/fixtures_ros2_adapter/rosidl_runtime_c/primitives_sequence_functions.h
@@ -1,0 +1,47 @@
+/*
+ * Copyright (c) 2025-2026 TSN Lab, Inc.
+ *
+ * SPDX-License-Identifier: GPL-3.0-or-later
+ *
+ * This file is part of TickLE. TickLE is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License, version 3, as published by the Free
+ * Software Foundation. A proprietary license is also available on request - see README.md.
+ */
+
+#pragma once
+#include <stdbool.h>
+#include <stddef.h>
+#include <stdint.h>
+#include <stdlib.h>
+
+typedef struct rosidl_runtime_c__uint16__Sequence {
+    uint16_t* data;
+    size_t size;
+    size_t capacity;
+} rosidl_runtime_c__uint16__Sequence;
+
+static inline bool rosidl_runtime_c__uint16__Sequence__init(rosidl_runtime_c__uint16__Sequence* seq, size_t size) {
+    seq->data = (uint16_t*)malloc(size * sizeof(uint16_t));
+    if (seq->data == NULL && size > 0) {
+        return false;
+    }
+    seq->size = size;
+    seq->capacity = size;
+    return true;
+}
+
+typedef struct rosidl_runtime_c__float32__Sequence {
+    float* data;
+    size_t size;
+    size_t capacity;
+} rosidl_runtime_c__float32__Sequence;
+
+static inline bool rosidl_runtime_c__float32__Sequence__init(rosidl_runtime_c__float32__Sequence* seq, size_t size) {
+    seq->data = (float*)malloc(size * sizeof(float));
+    if (seq->data == NULL && size > 0) {
+        return false;
+    }
+    seq->size = size;
+    seq->capacity = size;
+    return true;
+}